requirements.txt
hikari
hikari-lightbulb
uvloop
pandas
bs4